Move Charts and Table section to DataViz module
