+++ src/components/App.jsx
@@ -1,10 +1,9 @@
 import { styled } from 'linaria/react'
 import React from 'react'
 
-import { ChartSection } from './Layout/ChartSection'
 import { ConstraintSection } from './Layout/ConstraintSection'
+import { ChartSection, TableSection } from './Layout/DataVizSection'
 import { Header } from './Layout/Header'
-import { TableSection } from './Layout/TableSection'
 
 const StyledTableChartsSection = styled.section`
 	padding: 10px 30px 0;
