+++ src/components/Layout/DataVizSection.jsx
@@ -4,7 +4,18 @@ import React from 'react'
 
 import { humanMine25 as rows } from '../../stubs/humanMine25'
 import { mineUrl } from '../../stubs/utils'
-import { Table, TableActionButtons, TablePagingButtons } from '../DataViz'
+import { BarChart, PieChart, Table, TableActionButtons, TablePagingButtons } from '../DataViz'
+
+const S_Card = styled(Card)`
+	height: 376px;
+	margin-bottom: 20px;
+	display: flex;
+`
+
+const S_ChartContainer = styled.div`
+	height: 100%;
+	width: 45%;
+`
 
 const TableCard = styled(Card)`
 	margin-bottom: 20px;
@@ -25,10 +36,32 @@ const StyledPagingRow = styled.div`
 	justify-content: space-between;
 `
 
+const StyledTableChartsSection = styled.section`
+	padding: 10px 30px 0;
+	overflow: auto;
+	height: calc(100vh - 3.643em);
+`
+
 const S = {
 	TableCard,
 	RowCount: StyledRowCount,
 	PagingRow: StyledPagingRow,
+	TableChartSection: StyledTableChartsSection,
+}
+
+export const ChartSection = () => {
+	return (
+		<section id="charts">
+			<S_Card>
+				<S_ChartContainer>
+					<PieChart />
+				</S_ChartContainer>
+				<S_ChartContainer>
+					<BarChart />
+				</S_ChartContainer>
+			</S_Card>
+		</section>
+	)
 }
 
 export const TableSection = () => {
@@ -43,5 +76,14 @@ export const TableSection = () => {
 				<Table mineUrl={mineUrl} rows={rows} />
 			</S.TableCard>
 		</section>
+	)
+}
+
+export const TableChartSection = () => {
+	return (
+		<S.TableChartSection id="Tablechart">
+			<ChartSection />
+			<TableSection />
+		</S.TableChartSection>
 	)
 }
