+++ src/components/Layout/DataVizSection.story.jsx
@@ -1,10 +1,11 @@
 import React from 'react'
 
-import * as Charts from './ChartSection'
+import * as Charts from './DataVizSection'
 
 export default {
-	component: Charts.ChartSection,
 	title: 'Components/Data Visualization',
 }
 
 export const ChartSection = () => <Charts.ChartSection />
+
+export const TableSection = () => <Charts.TableSection />
