--- src/components/Layout/TableSection.story.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-import { TableSection } from './TableSection'
-
-export default {
-	component: TableSection,
-	title: 'Components/Table and Chart',
-}
-
-export const Table = () => <TableSection />
